% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelHIV-package.R
\docType{package}
\name{EpiModelHIV-package}
\alias{EpiModelHIV-package}
\alias{EpiModelHIV}
\title{Network-Based Epidemic Modeling of HIV/STI Transmission}
\description{
\tabular{ll}{
   Package: \tab EpiModelHIV\cr
   Type: \tab Package\cr
   Version: \tab 2.5.0\cr
   Date: \tab 2022-04-06\cr
   License: \tab GPL-3\cr
   LazyLoad: \tab yes\cr
}
}
\details{
EpiModelHIV provides extensions to our general EpiModel package to allow
for simulating HIV/STI transmission dynamics.
}
\keyword{het}
\keyword{msm}
\keyword{package}
