% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_msm}
\alias{param_msm}
\title{Epidemic Model Parameters}
\usage{
param_msm(
  netstats,
  hiv.test.rate = c(0.01325, 0.0125, 0.0124),
  hiv.test.late.prob = c(0, 0, 0),
  test.window.int = 3,
  tt.partial.supp.prob = c(0, 0, 0),
  tt.full.supp.prob = c(1, 1, 1),
  tt.durable.supp.prob = c(0, 0, 0),
  tx.init.rate = c(0.092, 0.092, 0.127),
  tx.halt.partial.rate = c(0.0102, 0.0102, 0.0071),
  tx.halt.full.rr = c(0.45, 0.45, 0.45),
  tx.halt.durable.rr = c(0.45, 0.45, 0.45),
  tx.reinit.partial.rate = c(0.00066, 0.00066, 0.00291),
  tx.reinit.full.rr = c(1, 1, 1),
  tx.reinit.durable.rr = c(1, 1, 1),
  max.time.off.tx.full.int = 780,
  max.time.on.tx.partial.int = 520,
  max.time.off.tx.partial.int = 520,
  vl.acute.rise.int = 3,
  vl.acute.peak = 6.886,
  vl.acute.fall.int = 3,
  vl.set.point = 4.5,
  vl.aids.onset.int = 520,
  vl.aids.int = 104,
  vl.aids.peak = 7,
  vl.full.supp = 1.5,
  vl.part.supp = 3.5,
  vl.tx.down.rate = 0.25,
  vl.tx.aids.down.rate = 0.25,
  vl.tx.up.rate = 0.25,
  aids.mort.rate = 0.009615,
  a.rate = 0.00052,
  arrival.age = 15,
  hiv.urai.prob = 0.008938,
  hiv.uiai.prob = 0.003379,
  hiv.trans.scale = c(1, 1, 1),
  hiv.acute.rr = 6,
  hiv.circ.rr = 0.4,
  hiv.cond.eff.rr = 0.95,
  hiv.cond.fail.rr = c(0.25, 0.25, 0.25),
  hiv.circ.prob = c(0.874, 0.874, 0.918),
  epistats,
  acts.aids.vl = 5.75,
  acts.scale = 1,
  cond.scale = 1,
  rgc.prob = 0.35,
  ugc.prob = 0.25,
  vgc.prob = 0.25,
  rct.prob = 0.2,
  uct.prob = 0.16,
  vct.prob = 0.16,
  rgc.sympt.prob = 0.16,
  ugc.sympt.prob = 0.8,
  vgc.sympt.prob = 0.8,
  rct.sympt.prob = 0.14,
  uct.sympt.prob = 0.58,
  vct.sympt.prob = 0.58,
  rgc.ntx.int = 16.8,
  ugc.ntx.int = 16.8,
  vgc.ntx.int = 16.8,
  gc.tx.int = 1.4,
  rct.ntx.int = 32,
  uct.ntx.int = 32,
  vct.ntx.int = 32,
  ct.tx.int = 1.4,
  gc.sympt.tx.prob = c(0.95, 0.95, 0.95, 0.95, 0.95, 0.95),
  ct.sympt.tx.prob = c(0.9, 0.9, 0.9, 0.9, 0.9, 0.9),
  gc.asympt.tx.prob = c(0.15, 0.15, 0.15, 0.15, 0.15, 0.15),
  ct.asympt.tx.prob = c(0.15, 0.15, 0.15, 0.15, 0.15, 0.15),
  sti.cond.eff.rr = 0.9,
  sti.cond.fail.rr = c(0.2, 0.2, 0.2),
  hiv.rgc.rr = 2.78,
  hiv.ugc.rr = 1.73,
  hiv.rct.rr = 2.78,
  hiv.uct.rr = 1.73,
  hiv.dual.rr = 0.2,
  riskh.start = Inf,
  prep.start = Inf,
  prep.start.prob = c(0.2, 0.2, 0.2),
  prep.adhr.dist = c(0.089, 0.127, 0.784),
  prep.adhr.rr = c(0.69, 0.19, 0.01),
  prep.discont.rate = rep(1 - (2^(-1/(224.4237/7))), 3),
  prep.tst.int = 12.8571,
  prep.risk.int = 26,
  prep.sti.screen.int = 26,
  prep.sti.tx.prob = 1,
  prep.risk.reassess.int = 52,
  prep.require.lnt = TRUE,
  part.ident.start = Inf,
  part.index.window.int = 0,
  part.index.degree = 1,
  part.index.prob = 1,
  part.ident.main.window.int = 12,
  part.ident.casl.window.int = 12,
  part.ident.ooff.window.int = 12,
  part.ident.main.prob = 1,
  part.ident.casl.prob = 1,
  part.ident.ooff.prob = 1,
  part.hiv.test.rate = c(1, 1, 1),
  part.prep.start.prob = c(0.5, 0.5, 0.5),
  part.tx.init.rate = c(0.6, 0.6, 0.8),
  part.tx.halt.rate = c(0.00102, 0.00102, 0.00071),
  part.tx.reinit.rate = c(0.5, 0.5, 0.5),
  ...
)
}
\arguments{
\item{netstats}{Target statistics and related network initialization data
from the standard ARTnet workflow.}

\item{hiv.test.rate}{Mean probability of HIV testing per time step for
Black/Hispanic/White MSM (vector of length 3).}

\item{hiv.test.late.prob}{Proportion of Black/Hispanic/White MSM who test
only during AIDS stage infection (vector of length 3).}

\item{test.window.int}{Length of the HIV test window period in time steps.}

\item{tt.partial.supp.prob}{Proportion of Black/Hispanic/White MSM who enter
partial viral suppression category after ART initiation (vector of
length 3).}

\item{tt.full.supp.prob}{Proportion of Black/Hispanic/White MSM who enter full
viral suppression category after ART initiation (vector of length 3).}

\item{tt.durable.supp.prob}{Proportion of Black/Hispanic/White MSM who enter
durable viral suppression category after ART initiation (vector of
length 3).}

\item{tx.init.rate}{Probability per time step that a Black/Hispanic/White MSM
who has tested positive will initiate treatment (vector of length 3).}

\item{tx.halt.partial.rate}{Probability per time step that
Black/Hispanic/White MSM who have started treatment and assigned to
the partial VL suppression category will stop treatment (vector of
length 3).}

\item{tx.halt.full.rr}{Relative reduction in \code{tx.halt.partial.rate} for
Black/Hispanic/White MSM in the full VL suppression category (vector
of length 3).}

\item{tx.halt.durable.rr}{Relative reduction in \code{tx.halt.partial.rate}
for Black/Hispanic/White MSM in the durable VL suppression category
(vector of length 3).}

\item{tx.reinit.partial.rate}{Probability per time step that a
Black/Hispanic/White MSM who has stopped treatment and assigned to the
partial VL suppression category will restart treatment (vector of
length 3).}

\item{tx.reinit.full.rr}{Relative reduction in \code{tx.reinit.partial.rate}
for Black/Hispanic/White MSM in the full VL suppression category
(vector of length 3).}

\item{tx.reinit.durable.rr}{Relative reduction in
\code{tx.reinit.partial.rate} for Black/Hispanic/White MSM in the
durable VL suppression category (vector of length 3).}

\item{max.time.off.tx.full.int}{Number of time steps off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.partial.int}{Number of time steps on treatment for a
partial suppressor before onset of AIDS.}

\item{max.time.off.tx.partial.int}{Number of time steps off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.int}{Number of time steps to peak viremia during acute
infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.int}{Number of time steps from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of time steps to AIDS for a treatment-naive
patient.}

\item{vl.aids.int}{Duration of AIDS stage infection in time steps.}

\item{vl.aids.peak}{Maximum viral load during AIDS stage.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{vl.tx.down.rate}{Number of log10 units that viral load falls per time
step from treatment initiation or re-initiation until the suppression
level is reached (pre-AIDS stages).}

\item{vl.tx.aids.down.rate}{Number of log10 units that viral load falls per
time step from treatment initiation or re-initiation until the
suppression level is reached (AIDS stage).}

\item{vl.tx.up.rate}{Number of log10 units that viral load rises per time
step from treatment halting until expected value.}

\item{aids.mort.rate}{Mortality rate of persons in the AIDS stage who are
currently off ART per time step.}

\item{a.rate}{Rate by time step at which MSM enter the population.}

\item{arrival.age}{Age (in years) of new arrivals.}

\item{hiv.urai.prob}{Probability of infection for a man having unprotected
receptive anal intercourse with an infected man at set point viral
load.}

\item{hiv.uiai.prob}{Probability of infection for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{hiv.trans.scale}{Relative scalar on base infection probabilities for model
calibration for Black/Hispanic/White men (vector of length 3).}

\item{hiv.acute.rr}{Relative risk of infection (compared to that predicted by
elevated viral load) when positive partner is in the acute stage.}

\item{hiv.circ.rr}{Relative risk of infection from insertive anal sex when the
negative insertive partner is circumcised.}

\item{hiv.cond.eff.rr}{Relative risk of HIV infection from anal sex when a condom is
used properly (biological efficacy).}

\item{hiv.cond.fail.rr}{Relative risk of condom failure per time step for HIV for
Black/Hispanic/White MSM, as a reduction in the cond.eff.rr parameter
(vector of length 3).}

\item{hiv.circ.prob}{Probability that a Black/Hispanic/White new arrival in the
population will be circumcised (vector of length 3).}

\item{epistats}{GLMs for epidemiological parameter from the standard ARTnet
workflow.}

\item{acts.aids.vl}{Viral load level after which sexual act rate goes to
zero.}

\item{acts.scale}{Scalar for main/casual act rate for model calibration.}

\item{cond.scale}{Scalar for condom use probability for model calibration.}

\item{rgc.prob}{Probability of rectal gonorrhea infection per act.}

\item{ugc.prob}{Probability of urethral gonorrhea infection per act.}

\item{vgc.prob}{Probability of vaginal gonorrhea infection per act.}

\item{rct.prob}{Probability of rectal chlamydia infection per act.}

\item{uct.prob}{Probability of urethral chlamydia infection per act.}

\item{vct.prob}{Probability of vaginal chlamydia infection per act.}

\item{rgc.sympt.prob}{Probability of symptoms given infection with rectal
gonorrhea.}

\item{ugc.sympt.prob}{Probability of symptoms given infection with urethral
gonorrhea.}

\item{vgc.sympt.prob}{Probability of symptoms given infection with vaginal
gonorrhea.}

\item{rct.sympt.prob}{Probability of symptoms given infection with rectal
chlamydia.}

\item{uct.sympt.prob}{Probability of symptoms given infection with urethral
chlamydia.}

\item{vct.sympt.prob}{Probability of symptoms given infection with vaginal
chlamydia.}

\item{rgc.ntx.int}{Average duration in time steps of untreated rectal gonorrhea.}

\item{ugc.ntx.int}{Average duration in time steps of untreated urethral gonorrhea.}

\item{vgc.ntx.int}{Average duration in time steps of untreated vaginal gonorrhea.}

\item{gc.tx.int}{Average duration in time steps of treated gonorrhea (all sites).}

\item{rct.ntx.int}{Average duration in time steps of untreated rectal chlamydia.}

\item{uct.ntx.int}{Average duration in time steps of untreated vaginal chlamydia.}

\item{ct.tx.int}{Average duration in time steps of treated chlamydia (all sites).}

\item{gc.sympt.tx.prob}{Probability of treatment for symptomatic gonorrhea
for Black/Hispanic/White men and women (vector of length 6).}

\item{ct.sympt.tx.prob}{Probability of treatment for symptomatic chlamydia
for Black/Hispanic/White men and women (vector of length 6).}

\item{gc.asympt.tx.prob}{Probability of treatment for asymptomatic gonorrhea
for Black/Hispanic/White men and women (vector of length 6).}

\item{ct.asympt.tx.prob}{Probability of treatment for asymptomatic chlamydia
for Black/Hispanic/White men and women (vector of length 6).}

\item{sti.cond.eff.rr}{Relative risk of STI infection from anal sex when a
condom is used properly (biological efficacy).}

\item{sti.cond.fail.rr}{Relative risk of condom failure for STI for
Black/Hispanic/White MSM, as a reduction in \code{sti.cond.eff.rr}
(vector of length 3).}

\item{hiv.rgc.rr}{Relative risk of HIV infection given current rectal
gonorrhea.}

\item{hiv.ugc.rr}{Relative risk of HIV infection given current urethral
gonorrhea.}

\item{hiv.rct.rr}{Relative risk of HIV infection given current rectal
chlamydia.}

\item{hiv.uct.rr}{Relative risk of HIV infection given current urethral
chlamydia.}

\item{hiv.dual.rr}{Additive proportional risk, from 0 to 1, for HIV infection
given dual infection with both gonorrhea and chlamydia.}

\item{riskh.start}{Time step at which behavioral risk history assessment
occurs.}

\item{prep.start}{Time step at which the PrEP intervention should start.}

\item{prep.start.prob}{Probability of starting PrEP given current
indications.}

\item{prep.adhr.dist}{Proportion of men who are low, medium, and high
adherent to PrEP.}

\item{prep.adhr.rr}{Relative risk for infection per act associated with
each level of adherence (from Grant).}

\item{prep.discont.rate}{Rate of random discontinuation from PrEP per time step.}

\item{prep.tst.int}{Testing interval by time steps for those who are actively on PrEP. This
overrides the mean testing interval parameters.}

\item{prep.risk.int}{Time window for assessment of risk eligibility for PrEP
in time steps.}

\item{prep.sti.screen.int}{Interval in time steps between STI screening at PrEP
visits.}

\item{prep.sti.tx.prob}{Probability of treatment given positive screening
during PrEP visit.}

\item{prep.risk.reassess.int}{Interval for reassessment of risk
indications of active PrEP users. Numerical value based on time unit
conversion to yearly reassessment of PrEP risk. Default of 52 for
yearly reassessment when time unit is weekly, 364 for yearly
reassessment when time unit is daily, 12 for yearly reassessment when
time unit is monthly, and Inf for no reassessment.}

\item{prep.require.lnt}{If \code{TRUE}, only start on PrEP if current time
step is equal to the last negative test.}

\item{part.ident.start}{Time step to initiate the partner identification
process. Default of \code{Inf} indicates no partner identification.}

\item{part.index.window.int}{Number of prior time steps for which a newly
diagnosed index patient is eligible for partner notification. Default
of 0 indicates index eligible only if newly diagnosed at current time
step.}

\item{part.index.degree}{The minimum number of partnerships an index patient
may have for partner elicitation. Default of \code{1} indicating all
partnerships under consideration.}

\item{part.index.prob}{Probability that an index patient will initiate
partner services.}

\item{part.ident.main.window.int}{The number of time steps in the past that a main
partnership qualifies for partner identification.}

\item{part.ident.casl.window.int}{The number of time steps in the past that a
casual partnership qualifies for partner identification.}

\item{part.ident.ooff.window.int}{The number of time steps in the past that a
one-off partnership qualifies for partner identification.}

\item{part.ident.main.prob}{Probability that a main partner is identified
through partner identification.}

\item{part.ident.casl.prob}{Probability that a casual partner is identified
through partner identification.}

\item{part.ident.ooff.prob}{Probability that a one-time partner is identified
through partner identification.}

\item{part.hiv.test.rate}{Mean probability per time step of HIV testing for a
Black/Hispanic/White MSM who has been identified as the partner of an
index HIV+ MSM.}

\item{part.prep.start.prob}{Probability of an individual identified through
partner identification starting PrEP given current indications at the
current time step.}

\item{part.tx.init.rate}{Probability that a Black/Hispanic/White MSM who has
been identified as the partner of an incident HIV+ MSM will initiate
treatment during the current time step.}

\item{part.tx.halt.rate}{Probability per time step that Black/Hispanic/White
MSM who have have been identified through partner identification,
started treatment and assigned to the partial VL suppression category
will stop treatment (vector of length 3).}

\item{part.tx.reinit.rate}{Probability per time step that a
Black/Hispanic/White MSM who has been identified through partner
identification, stopped treatment will restart treatment (vector of
length 3).}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{param_msm}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for EpiModelHIV
}
\section{Modifying Time-Specific Parameters}{

The default time steps of these parameters are weekly time steps.
Users must modify parameters ending in .rate, .int, and .start when using
non-weekly time steps. For example, when using daily time steps, .int
parameters should be multiplied by 7 and .rate parameters should be divided
by 7.
}

\keyword{msm}
