% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.partident.R
\name{partident_msm}
\alias{partident_msm}
\title{Partner Identification Module}
\usage{
partident_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with at updated \code{part.ident}
attribute detailing time of identification.
}
\description{
Module function for identifying partners of incident HIV+ MSM.
}
\details{
This partner identification module handles the elicitation of partners of
incident HIV+ MSM for the purposes of HIV testing (\link[EpiModelHIV]{hivtest_msm}).
}
\keyword{module}
\keyword{msm}
